% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_add_col_attributes.R
\name{cb_add_col_attributes}
\alias{cb_add_col_attributes}
\title{Add Attributes to Columns}
\usage{
cb_add_col_attributes(df, .x, ...)
}
\arguments{
\item{df}{Data frame of interest}

\item{.x}{Column of interest in df}

\item{...}{Arbitrary list of attributes (i.e., attribute = "value")}
}
\value{
Data frame
}
\description{
Add arbitrary attributes to columns (e.g., description, source, column type).
These attributes can later be accessed to fill in the column attributes table.
Note: Column type is the type I say it is: categorical or numeric, Data type
is what the computer says it is.
}
