% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_add_section_header.R
\name{cb_add_section_header}
\alias{cb_add_section_header}
\title{Create Formatted Section Header}
\usage{
cb_add_section_header(rdocx, text = NA)
}
\arguments{
\item{rdocx}{rdocx object created with \code{officer::read_docx()}}

\item{text}{Text of section header}
}
\value{
rdocx object
}
\description{
Create Formatted Section Header
}
\keyword{internal}
