% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_summary_stats_to_ft.R
\name{cb_summary_stats_to_ft}
\alias{cb_summary_stats_to_ft}
\alias{cb_summary_stats_to_ft.summary_numeric}
\alias{cb_summary_stats_to_ft.summary_many_cats}
\alias{cb_summary_stats_to_ft.summary_few_cats}
\alias{cb_summary_stats_to_ft.summary_time}
\title{Create Formatted Flextable From Summary Statistics}
\usage{
cb_summary_stats_to_ft(df, ...)

\method{cb_summary_stats_to_ft}{summary_numeric}(df, col_width = 1.3)

\method{cb_summary_stats_to_ft}{summary_many_cats}(df, col_width = 1.62)

\method{cb_summary_stats_to_ft}{summary_few_cats}(df, col_width = 1.62)

\method{cb_summary_stats_to_ft}{summary_time}(df, col_width = 1.62)
}
\arguments{
\item{df}{Data frame of summary statistics}

\item{...}{Other stuff}

\item{col_width}{Set the width of the column that will appear in the Word table}
}
\value{
Flextable object
}
\description{
Create Formatted Flextable From Summary Statistics
}
\keyword{internal}
