% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_theme_col_attr.R
\name{cb_theme_col_attr}
\alias{cb_theme_col_attr}
\title{Format Column Attributes Flextable}
\usage{
cb_theme_col_attr(ft)
}
\arguments{
\item{ft}{A flextable object}
}
\value{
A flextable object
}
\description{
Format Column Attributes Flextable
}
