% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_theme_df_attributes.R
\name{cb_theme_df_attributes}
\alias{cb_theme_df_attributes}
\title{Format Data Frame Attributes Flextable}
\usage{
cb_theme_df_attributes(ft)
}
\arguments{
\item{ft}{A flextable object}
}
\value{
A flextable object
}
\description{
Format Data Frame Attributes Flextable
}
\keyword{internal}
