% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Automate creation of a data codebook}
\usage{
codebook(df, path = NA, title = NA, subtitle = NA, description = NA)
}
\arguments{
\item{df}{The saved file at \code{path}, read into memory as a data frame}

\item{path}{The path to the saved dataset of interest}

\item{title}{Optional title}

\item{subtitle}{Optional subtitle}

\item{description}{Text description of the dataset}
}
\value{
An rdocx object that can be printed to a Word document
}
\description{
The codebook function assists with the creation of a codebook for a given
data frame.
}
\details{
Codebook expects that \code{df } is a data frame that you have read into memory
from a saved data file. Please provide the path to the saved data file. This
function gets selected attributes about file saved at \code{path} and stores
those attributes in a data frame, which is later turned into a flextable and
added to the codebook document.
}
\examples{
# codebook_detect_5wk <- codebook(
#   df = detect_5wk \%>\% select(1:2),
#   path = "../data/detect_5wk.csv",
#   title = "Detection of Elder abuse Through Emergency Care Technicians (DETECT)",
#   subtitle = "5-Week Pilot Study",
#   description = description
# ) \%>\%
#   print(target = "example_officer_codebook.docx")
}
