% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_add_description.R
\name{codebook_add_description}
\alias{codebook_add_description}
\title{Add Description Text to Codebook}
\usage{
codebook_add_description(rdocx, description)
}
\arguments{
\item{rdocx}{rdocx rdocx object created with \code{officer::read_docx()}}

\item{description}{Text description of the dataset}
}
\value{
rdocx object
}
\description{
Basically, just checks for the number of paragraphs in the description and
then runs codebook_add_text for each one.
}
