% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_add_summary_stats.R
\name{codebook_add_summary_stats}
\alias{codebook_add_summary_stats}
\title{Calculate Appropriate Statistics for Variable Type}
\usage{
codebook_add_summary_stats(
  df,
  .x,
  many_cats = 10,
  num_to_cat = 4,
  digits = 2,
  n_extreme_cats = 5
)
}
\arguments{
\item{df}{Data frame of interest}

\item{.x}{Column of interest}

\item{many_cats}{Set cutoff value for many categories/few categories}

\item{num_to_cat}{Set cutoff value for guessing that a numeric variable is actually categorical}

\item{digits}{Number of digits after decimal to display}

\item{n_extreme_cats}{Number of extreme values to display}
}
\value{
A tibble of results
}
\description{
Calculate Appropriate Statistics for Variable Type
}
