% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_add_text.R
\name{codebook_add_text}
\alias{codebook_add_text}
\title{Create Formatted Text}
\usage{
codebook_add_text(rdocx, text = NA)
}
\arguments{
\item{rdocx}{rdocx rdocx object created with \code{officer::read_docx()}}

\item{text}{Arbitrary text}
}
\value{
rdocx object
}
\description{
Create Formatted Text
}
