% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_get_col_attributes.R
\name{codebook_get_col_attributes}
\alias{codebook_get_col_attributes}
\title{Get Column Attributes}
\usage{
codebook_get_col_attributes(df, .x)
}
\arguments{
\item{df}{Data frame of interest}

\item{.x}{Column of interest in df}
}
\value{
A tibble
}
\description{
Get Column Attributes
}
