% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_summary_stats_many_cats.R
\name{codebook_summary_stats_many_cats}
\alias{codebook_summary_stats_many_cats}
\title{Compute Summary Statistics for Categorical Variables with Many Categories}
\usage{
codebook_summary_stats_many_cats(df, .x, n_extreme_cats = 5)
}
\arguments{
\item{df}{Data frame of interest}

\item{.x}{Column of interest}

\item{n_extreme_cats}{Number of extreme values to display}
}
\value{
A tibble
}
\description{
Compute Summary Statistics for Categorical Variables with Many Categories
}
