% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_summary_stats_numeric.R
\name{codebook_summary_stats_numeric}
\alias{codebook_summary_stats_numeric}
\title{Compute Summary Statistics for Numeric Variables}
\usage{
codebook_summary_stats_numeric(df, .x, digits = 2)
}
\arguments{
\item{df}{Data frame of interest}

\item{.x}{Column of interest}

\item{digits}{Number of digits after decimal to display}
}
\value{
A tibble
}
\description{
Compute Summary Statistics for Numeric Variables
}
