% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_summary_stats_time.R
\name{codebook_summary_stats_time}
\alias{codebook_summary_stats_time}
\title{Compute Summary Statistics for Time Variables}
\usage{
codebook_summary_stats_time(df, .x, digits = 2)
}
\arguments{
\item{df}{Data frame of interest}

\item{.x}{Column of interest}

\item{digits}{Number of digits after decimal to display}
}
\value{
A tibble
}
\description{
Compute Summary Statistics for Time Variables
}
