% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebookr-package.R
\docType{package}
\name{codebookr}
\alias{codebookr}
\alias{codebookr-package}
\title{codebookr: A package for reading codebooks and applying to datasets}
\description{
The codebookr package provides functions for reading a codebook from a
spreadsheet (.csv file), creating factors and checking levels and
checking continuous values are in a valid range.
}
