% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook.R
\name{create_codebook}
\alias{create_codebook}
\title{Create a PDF codebook}
\usage{
create_codebook(
  file_path,
  datasets_input,
  variables_input,
  title_text,
  version_text,
  footer_text,
  author_names,
  table_of_contents = TRUE,
  include_variable_type = FALSE,
  theme_color = "#3B86F7",
  title_font_size = 16,
  heading_font_size = 35,
  subheading_font_size = 12
)
}
\arguments{
\item{file_path}{The path and file for the \code{.tex} file that will be
created. You need to include the \code{.tex} extension in the file name.}

\item{datasets_input}{A data frame containing information on each dataset.
This data frame must include: (1) a variable called \code{dataset} that is
the name of each dataset, which will be used as the heading for each
dataset section; (2) a variable called \code{label} that is a short label
for each dataset (one line), which will be used as the subheading for each
dataset section; and (3) a variable called \code{description} that is a
description of each dataset, which will be included before the description
of each variable. Text wrapped in braces will be formatted as code. The
dataset names should be valid \code{R} object names.}

\item{variables_input}{A data frame containing infromation on each variable.
This data frame must include: (1) a variable called \code{dataset}, where
each dataset maps exactly to the datasets in the \code{dataset} variable in
the \code{datasets_input} data frame; and (2) a variable called
\code{description} that is a description of each variable. Text wrapped in
braces will be formatted as code. Optionally, the data frame can include a
variable called \code{type} that indicates the type of each variable (e.g.,
numeric, string, etc.). If this variable is included, and the
\code{include_variable_type} argument is set to \code{TRUE}, then the
variable type will be included at the beginning of each variable
description.}

\item{title_text}{A string containing the title for the title page.}

\item{version_text}{A string containing the version number for the title page
(e.g., \code{1.0}).}

\item{author_names}{A string vector containing the names of the
authors for the title page or a string if there is one author.}

\item{table_of_contents}{A logical value indicating whether to include a
table of contents.}

\item{include_variable_type}{A logical value indicating whether to include
the type of the variable in the variable description. The
\code{variables_input} data frame must have a variable called \code{type}
or you will get an error.}

\item{theme_color}{A string indicating the color to use. The color should be
a valid hex code, including a leading \code{#}. If you don't provide a
valid hex code, your \code{LaTeX} compiler will produce an error (but this
function will not).}

\item{title_font_size}{The size of the font for the title on the title page.
You should adjust the font size to make sure your text fits.}

\item{heading_font_size}{The size of the font for the heading for each
dataset section (i.e., the name of the dataset, contained in the
\code{dataset} variable in the \code{datasets_input} data frame). You
should adjust the font size to make sure your text fits.}

\item{subheading_font_size}{The size of the font for the subheading for each
dataset section (i.e., the short label for the dataset, contained in the
\code{label} variable in the \code{datasets_input} data frame). You should
adjust the font size to make sure your text fits.}
}
\description{
This function generates \code{LaTeX} code (a \code{.tex} file) for a
nicely-formatted PDF codebook. You need to provide a data frame with a
description of each dataset and a data frame with a description of each
variable. Once you've generated the \code{.tex} file, you can compile the PDF
using any \code{LaTeX} editor. You need to compile the \code{.tex} file using
\code{XeLaTeX} instead of \code{LaTeX}. You will need to compile the PDF
twice. You need to have the \code{roboto} font installed, which you can
download for free on Google Fonts
(\code{https://fonts.google.com/specimen/Roboto}). You can further customize
the \code{.tex} file that is generated. Please report bugs and request
features at \code{https://github.com/jfjelstul/codebookr/issues}.
}
