% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{create_factors}
\alias{create_factors}
\title{Use codebook to create factors and check levels for validity}
\usage{
create_factors(x, code_book, column_names = NULL)
}
\arguments{
\item{x}{\code{tibble} to which \code{codebook} is applied}

\item{code_book}{\code{codebook} containing factor names and factor levels
to convert numeric or character vectors to \code{factors} and also test
for valid levels.}

\item{column_names}{character vector of column names for conversion to
factors. Default: All \code{factors} defined in \code{codebook}.}
}
\value{
object of type class \code{tibble}
}
\description{
Uses a codebook which is an S3 class \code{codebook}, possibly
read in by \code{read_codebook}, to convert numerical vectors into
\code{factors} in a \code{tibble}.
}
\details{
REPEAT: Often, when analysing data, data dictionaries or code books are
provided with data files. Rather than a \code{word} \code{doc} or
\code{pdf} files, the format required here is in a very specific
format stored as a \code{csv} file. Once read in, attributes such
as factor labels/levels and variable labels can be added to the
\code{data.frame} and/or also used to check factor labels and
variable names are consistent with the code book. Note that while
various methods may be available which attempt to convert word
docs or pdf's to a spreadsheet and/or csv file, extreme care
should be taken as these are far from perfect.
}
\examples{
file.copy(system.file('demoFiles', 'data1_codebook.csv',
                      package='codebookr'), 'data1_codebook.csv')
file.copy(system.file('demoFiles', 'data1-birth.csv',
                      package='codebookr'), 'data1-birth.csv')
data1_codebook <- read_codebook("data1_codebook.csv",
         column_names = list(variable_levels = "Factor.Levels",
                             variable_original = "Old.Variable",
                             min = "Min", max = "Max"))
data1 <- readr::read_csv('data1-birth.csv')
data1
myData <- create_factors(data1,  data1_codebook)
str(myData)
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}
