% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.codebook.R
\name{print.codebook}
\alias{print.codebook}
\alias{codebook}
\title{Print an S3 object of class \code{codebook}}
\usage{
\method{print}{codebook}(x, extra = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{codebook}}

\item{extra}{logical: whether to print extra information. Default: FALSE}

\item{...}{extra arguments passed to specific printing functions}
}
\description{
\code{read_codebook} reads a code book stored as a \code{csv} file
for either checking against a data file or relabelling factor
levels or labelling variables.  \code{read_codebook} returns an S3
object of  class \code{codebook}
}
\examples{
file.copy(system.file('demoFiles', 'data1_codebook.csv',
                      package='codebookr'), 'data1_codebook.csv')
data1_codebook <- read_codebook("data1_codebook.csv",
         column_names = list(variable_levels = "Factor.Levels",
                             variable_original = "Old.Variable",
                             min = "Min", max = "Max"))
print(data1_codebook)
}
\seealso{
\code{\link{read_codebook}}
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}
