% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_codebook.R
\name{read_codebook}
\alias{read_codebook}
\title{Read a code book in standard format(s) as a csv file}
\usage{
read_codebook(x, codebook_directory = NULL, column_names = NULL,
  na = c("", "NA", ".", " "),
  data_management_plan = "All possible details should be here")
}
\arguments{
\item{x}{filename of codebook to parse}

\item{codebook_directory}{directory containing codebook. Default :
current directory}

\item{column_names}{named character vector containing column names
in Code Book file. The vector contains components
\code{variable_name} = variable name,
\code{variable_original} = original name (if variable_name was
changed), \code{label} for printing/plotting,
\code{variable_levels} = factor levels, \code{variable_limits} or
\code{min} and \code{max} for continuous measurements,
\code{missing_values} = numeric or strings for values of
variable to be set as missing \code{comments} = comments about
the variable which may include the measurement instrument or
references about the measurement. Note that default values may
be found with \code{options(codebookr.column_names)}}

\item{na}{a character vector of strings which are to be
interpreted as \sQuote{NA} values.  Blank fields are also
considered to be missing values in logical, integer, numeric and
complex fields. Default: \code{c("", "NA", ".", " ")}}

\item{data_management_plan}{a list containing information like url,
location, authors, date, version and so on.
Default: "All possible details should be here"}
}
\value{
S3 object of type class \code{codebook}
}
\description{
\code{read_codebook} reads a code book stored as a \code{csv} file
for either checking against a data file or relabelling factor
levels or labelling variables.
}
\details{
Often, when analysing data, data dictionaries or code books are
provided with data files. Rather than a \code{word} \code{doc} or
\code{pdf} files, the format required here is in a very specific
format stored as a \code{csv} file. Once read in, attributes such
as factor labels/levels and variable labels can be added to the
\code{data.frame} and/or also used to check factor labels and
variable names are consistent with the code book. Note that while
various methods may be available which attempt to convert word
docs or pdf's to a spreadsheet and/or csv file, extreme care
should be taken as these are far from perfect.
}
\examples{
file.copy(system.file('demoFiles', 'data1_codebook.csv',
                      package='codebookr'), 'data1_codebook.csv')
data1_codebook <- read_codebook("data1_codebook.csv",
         column_names = list(variable_levels = "Factor.Levels",
                             variable_original = "Old.Variable",
                             min = "Min", max = "Max"))
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}
