% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study}
\alias{study}
\title{Simulated study data.}
\format{
A data frame with 20 rows and 8 variables:
\describe{
\item{id}{Participant id}
\item{sex}{Participant sex, female/male}
\item{date}{Date of enrollment}
\item{time}{Time of enrollemnt}
\item{days}{Number of days enrolled}
\item{height}{Participant height, in inches}
\item{likert}{Likert scale item, 1-5}
\item{outcome}{Participant experienced the outcome of interest, TRUE or FALSE}
}
}
\usage{
study
}
\description{
This is the code to create the study data - a simulated dataset that can be
used to demonstrate how to use the codebook package.
}
\keyword{datasets}
