% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_limits.R
\name{validate_limits}
\alias{validate_limits}
\title{Use codebook to check continuous variables for validity}
\usage{
validate_limits(x, code_book, column_names = NULL)
}
\arguments{
\item{x}{\code{tibble} to which \code{codebook} is applied}

\item{code_book}{\code{codebook} containing names of continuous
variables and limits for checking data are within range}

\item{column_names}{character vector of column names for checking
data in range. Default: All continuous variables defined in
\code{codebook}.}
}
\value{
object of type class \code{tibble} containing data out of
  limits
}
\description{
Uses a codebook which is an S3 class \code{codebook}, possibly
read in by \code{read_codebook}, to check for data within valid
ranges.
}
\details{
REPEAT: Often, when analysing data, data dictionaries or code books are
provided with data files. Rather than a \code{word} \code{doc} or
\code{pdf} files, the format required here is in a very specific
format stored as a \code{csv} file. Once read in, attributes such
as factor labels/levels and variable labels can be added to the
\code{data.frame} and/or also used to check factor labels and
variable names are consistent with the code book. Note that while
various methods may be available which attempt to convert word
docs or pdf's to a spreadsheet and/or csv file, extreme care
should be taken as these are far from perfect.
}
\examples{
file.copy(system.file('demoFiles', 'data1_codebook.csv',
                      package='codebookr'), 'data1_codebook.csv')
file.copy(system.file('demoFiles', 'data1-yr21.csv',
                      package='codebookr'), 'data1-yr21.csv')
data1_codebook <- read_codebook("data1_codebook.csv",
         column_names = list(variable_levels = "Factor.Levels",
                             variable_original = "Old.Variable",
                             min = "Min", max = "Max"))
data1 <- readr::read_csv('data1-yr21.csv')
data1
non_valid <- validate_limits(data1, data1_codebook)
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}
