% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_summary_stats_few_cats.R
\name{cb_summary_stats_few_cats}
\alias{cb_summary_stats_few_cats}
\title{Compute Summary Statistics for Categorical Variables with Few Categories}
\usage{
cb_summary_stats_few_cats(df, .x, digits = 2)
}
\arguments{
\item{df}{Data frame of interest}

\item{.x}{Column of interest}

\item{digits}{Number of digits after decimal to display}
}
\value{
A tibble
}
\description{
Compute Summary Statistics for Categorical Variables with Few Categories
}
