% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study}
\alias{study}
\title{Simulated study data.}
\format{
A data frame with 20 rows and 4 variables:
\describe{
\item{id}{participant id}
\item{gender}{participant gender, female/male}
\item{date}{date of enrollment}
\item{height}{participant height, in inches}
}
}
\usage{
study
}
\description{
This is the code to create the study data - a simulated dataset that can be
used to demonstrate how to use the codebook package.
}
\keyword{datasets}
