% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study}
\alias{study}
\title{Simulated study data.}
\format{
A data frame with 20 rows and 8 variables:
\describe{
\item{id}{Participant's study identification number}
\item{sex}{Biological sex of the participant assigned at birth, female/male}
\item{date}{Participant's date of enrollment}
\item{time}{Participant's time of enrollment}
\item{days}{Total number of days the participant was enrolled in the study}
\item{height}{Participant's height in inches at date of enrollment}
\item{likert}{An example Likert scale item, 1-5}
\item{outcome}{Participant experienced the outcome of interest, TRUE or FALSE}
}
}
\usage{
study
}
\description{
This is the code to create the study data - a simulated dataset that can be
used to demonstrate how to use the codebook package.
}
\keyword{datasets}
